Refactor: move editor definition to staking widget
