+++ src/modules/dashboard/components/Dialogs/RaiseObjectionDialog/RaiseObjectionDialog.tsx
@@ -1,11 +1,4 @@
 import React, { useCallback, RefObject } from 'react';
-import { useEditor } from '@tiptap/react';
-import CharacterCount from '@tiptap/extension-character-count';
-import Color from '@tiptap/extension-color';
-import Placeholder from '@tiptap/extension-placeholder';
-import TextStyle from '@tiptap/extension-text-style';
-import Underline from '@tiptap/extension-underline';
-import StarterKit from '@tiptap/starter-kit';
 import { FormikProps } from 'formik';
 import * as yup from 'yup';
 import { bigNumberify } from 'ethers/utils';
@@ -34,8 +27,6 @@ interface Props extends FormProps {
 
 const displayName = 'dashboard.RaiseObjectionDialog';
 
-const characterLimit = 4000;
-
 const RaiseObjectionDialog = ({
   cancel,
   close,
@@ -46,20 +37,6 @@ const RaiseObjectionDialog = ({
   scrollToRef,
   ...props
 }: Props) => {
-  const editor = useEditor({
-    extensions: [
-      StarterKit,
-      Underline,
-      TextStyle,
-      Color,
-      CharacterCount.configure({ limit: characterLimit }),
-      Placeholder.configure({
-        emptyEditorClass: 'is-editor-empty',
-        placeholder: 'Enter the description...',
-      }),
-    ],
-  });
-
   const { walletAddress } = useLoggedInUser();
 
   const validationSchema = yup.object().shape({
@@ -124,7 +101,6 @@ const RaiseObjectionDialog = ({
             minUserStake={minUserStake}
             cancel={cancel}
             {...props}
-            editor={editor}
           />
         </Dialog>
       )}
